requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
backtrader>=1.9.78
matplotlib>=3.7.0
plotly>=5.15.0
scipy>=1.10.0
jinja2>=3.1.0
pyyaml>=6.0
